Casted to int

+++ STH/Assets/Scripts/GameManager.cs
@@ -552,14 +552,14 @@ public class GameManager : MonoBehaviour
 
             GameObject ghost = Instantiate(prefabToSpawn, candidatePos, Quaternion.identity, obstacleHolder);
 
-            ghost.transform.localScale = (randomIndex == Obstacles.BUILDING)
+            ghost.transform.localScale = (randomIndex == (int)Obstacles.BUILDING)
                 ? new Vector3(Random.Range(2, 10), Random.Range(5, 10), Random.Range(2, 10))
                 : Vector3.one;
 
             Quaternion rotation;
-            if (randomIndex == Obstacles.BUILDING)
+            if (randomIndex == (int)Obstacles.BUILDING)
                 rotation = Quaternion.identity;
-            else if (randomIndex == Obstacles.TREE || randomIndex == Obstacles.TENT)
+            else if (randomIndex == (int)Obstacles.TREE || randomIndex == (int)Obstacles.TENT)
                 rotation = Quaternion.Euler(0f, Random.Range(0f, 360f), 0f);
             else
                 rotation = Quaternion.identity;
